requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==3.2.4
django-ckeditor==6.5.1
django-embed-video==1.4.8
django-extra-views==0.14.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
spatialite==0.0.3
sqlparse==0.4.3
typing_extensions==4.6.0
urllib3==2.0.2
